:truck: refactor(`ui/settings/hooks`): Move `useSync` to its own file.
