+++ imports/ui/settings/hooks.ts
@@ -1,13 +1,4 @@
-import {
-	startTransition,
-	useCallback,
-	useEffect,
-	useMemo,
-	useState,
-} from 'react';
-
-import debounce from 'p-debounce';
-import {useDebounce} from 'use-debounce';
+import {useCallback, useMemo} from 'react';
 
 import {type SettingDocument, Settings} from '../../api/collection/settings';
 import {type UserSettings, type SettingKey, defaults} from '../../api/settings';
@@ -24,6 +15,7 @@ import {
 	TIMEOUT_INPUT_DEBOUNCE,
 	TIMEOUT_REACTIVITY_DEBOUNCE,
 } from '../constants';
+import {useSync} from '../../ux/hooks/useSync';
 
 const useSettingSubscription = <K extends SettingKey>(key: K) =>
 	useSubscription(byKey, [key]);
@@ -210,77 +202,6 @@ export const useSettingDebounced = <K extends SettingKey>(
 		value: clientValue,
 		setValue,
 		resetValue,
-	};
-};
-
-export const useSync = <T>(
-	loading: boolean,
-	serverValue: T,
-	inputDebounceTimeout: number,
-	reactivityDebounceTimeout: number,
-) => {
-	const [debouncedServerValue, {isPending, flush}] = useDebounce(
-		serverValue,
-		reactivityDebounceTimeout,
-		// NOTE: This should really be `{leading: !loading}`, but that does
-		// seem to work as `isPending` returns `true` even when leading update
-		// has been taken into account. See `flush`-based workaround below.
-		// SEE: https://github.com/xnimorz/use-debounce/issues/192
-		{leading: false},
-	);
-	useEffect(() => {
-		// NOTE: Flush when loading is done.
-		if (!loading) flush();
-	}, [loading, flush]);
-
-	const [clientValue, setClientValue] = useState(serverValue);
-	const [ignoreServer, setIgnoreServer] = useState(false);
-
-	useEffect(() => {
-		if (ignoreServer) return;
-		setClientValue((prev) => (isPending() ? prev : debouncedServerValue));
-	}, [ignoreServer, isPending, debouncedServerValue]);
-
-	const sync = useMemo(() => {
-		let last = {};
-
-		const debouncedSetServerValue = debounce(
-			async (current: unknown, setServerValue: () => Promise<void>) => {
-				try {
-					await setServerValue();
-				} finally {
-					setTimeout(() => {
-						startTransition(() => {
-							flush(); // NOTE: Fast-forward debounced state to current DB state.
-							// NOTE: Continue ignoring updates if we are not the
-							// last pending call. This works because all method
-							// calls are queued in calling order.
-							setIgnoreServer((prev) => (last === current ? false : prev));
-						});
-					}, inputDebounceTimeout + reactivityDebounceTimeout);
-				}
-			},
-			inputDebounceTimeout,
-		);
-
-		return async (
-			setClientValue: () => void,
-			setServerValue: () => Promise<void>,
-		) => {
-			setIgnoreServer(true); // NOTE: We ignore all updates.
-			const current = {};
-			last = current;
-			setClientValue();
-			await debouncedSetServerValue(current, setServerValue);
-			// NOTE: We will listen to updates again some time after last update
-			// is complete.
-		};
-	}, [setIgnoreServer, flush, inputDebounceTimeout, reactivityDebounceTimeout]);
-
-	return {
-		value: clientValue,
-		setValue: setClientValue,
-		sync,
 	};
 };
